+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 1.2.0 [in progress]
 ### Features
 1. [#120](https://github.com/influxdata/influxdb-client-go/pull/120) Health check API   
+1. [#121](https://github.com/influxdata/influxdb-client-go/pull/121) Remove trailing slash from connection URL  
 
 ### Breaking Change
  - [#107](https://github.com/influxdata/influxdb-client-go/pull/100) Renamed `InfluxDBClient` interface to `Client`, so the full name `influxdb2.Client` suits better to Go naming conventions